Moved order detail product list adapter to it's own class
