+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/orders/OrderDetailProductListView.kt
@@ -2,21 +2,15 @@ package com.woocommerce.android.ui.orders
 
 import android.content.Context
 import android.util.AttributeSet
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import androidx.recyclerview.widget.DividerItemDecoration
-import androidx.recyclerview.widget.RecyclerView
 import com.google.android.material.card.MaterialCardView
 import com.woocommerce.android.R
 import com.woocommerce.android.analytics.AnalyticsTracker
 import com.woocommerce.android.analytics.AnalyticsTracker.Stat
-import com.woocommerce.android.analytics.AnalyticsTracker.Stat.ORDER_DETAIL_PRODUCT_TAPPED
-import com.woocommerce.android.model.Order
 import com.woocommerce.android.model.Refund
 import com.woocommerce.android.model.toAppModel
 import com.woocommerce.android.tools.ProductImageMap
-import com.woocommerce.android.ui.products.ProductHelper
 import com.woocommerce.android.widgets.AlignedDividerDecoration
 import kotlinx.android.synthetic.main.order_detail_product_list.view.*
 import org.wordpress.android.fluxc.model.WCOrderModel
@@ -32,7 +26,7 @@ class OrderDetailProductListView @JvmOverloads constructor(
     init {
         View.inflate(context, R.layout.order_detail_product_list, this)
     }
-    private lateinit var viewAdapter: ProductListAdapter
+    private lateinit var viewAdapter: OrderDetailProductListAdapter
     private var isExpanded = false
 
     /**
@@ -76,7 +70,7 @@ class OrderDetailProductListView @JvmOverloads constructor(
                     )
                 }
 
-        viewAdapter = ProductListAdapter(
+        viewAdapter = OrderDetailProductListAdapter(
                 filteredItems,
                 productImageMap,
                 formatCurrencyForDisplay,
@@ -152,35 +146,5 @@ class OrderDetailProductListView @JvmOverloads constructor(
         productList_btnDetails.setOnClickListener(null)
         productList_btnFulfill.visibility = View.GONE
         productList_btnDetails.visibility = View.GONE
-    }
-
-    class ProductListAdapter(
-        private val orderItems: List<Order.Item>,
-        private val productImageMap: ProductImageMap,
-        private val formatCurrencyForDisplay: (BigDecimal) -> String,
-        private var isExpanded: Boolean,
-        private val productListener: OrderProductActionListener?
-    ) : RecyclerView.Adapter<ProductListAdapter.ViewHolder>() {
-        class ViewHolder(val view: OrderDetailProductItemView) : RecyclerView.ViewHolder(view)
-
-        override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-            val view: OrderDetailProductItemView = LayoutInflater.from(parent.context)
-                    .inflate(R.layout.order_detail_product_list_item, parent, false)
-                    as OrderDetailProductItemView
-            return ViewHolder(view)
-        }
-
-        override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-            val item = orderItems[position]
-            val productId = ProductHelper.productOrVariationId(item.productId, item.variationId)
-            val productImage = productImageMap.get(productId)
-            holder.view.initView(orderItems[position], productImage, isExpanded, formatCurrencyForDisplay)
-            holder.view.setOnClickListener {
-                AnalyticsTracker.track(ORDER_DETAIL_PRODUCT_TAPPED)
-                productListener?.openOrderProductDetail(productId)
-            }
-        }
-
-        override fun getItemCount() = orderItems.size
     }
 }
